add `dynamics training` docs & restructure docs
