+++ .gitignore
@@ -4,6 +4,7 @@ publishment.md
 
 book/
 docs/examples
+docs/apis/jaxsetting.rst
 others
 examples/recurrent_neural_network/neurogym
 develop/iconip_paper
