--- docs/apis/running.rst
@@ -1,12 +1,0 @@
-brainpy.running
-===============
-
-.. currentmodule:: brainpy.running
-.. automodule:: brainpy.running
-
-.. autosummary::
-    :toctree: generated/
-
-    process_pool
-    process_pool_lock
-
