+++ docs/tutorial_training/index.rst
@@ -5,3 +5,4 @@ Dynamics Training
    :maxdepth: 1
 
    build_artificial_network
+   creat_custom_layers
